vinvoor: refactor the cards page
