+++ vinvoor/src/components/LoadingSkeleton.tsx
@@ -1,0 +1,19 @@
+import { Skeleton, SkeletonProps } from "@mui/material";
+import { FC, ReactNode } from "react";
+
+interface LoadingSkeletonProps extends SkeletonProps {
+    loading: boolean;
+    children: ReactNode;
+}
+
+export const LoadingSkeleton: FC<LoadingSkeletonProps> = ({
+    loading,
+    children,
+    ...props
+}) => {
+    return loading ? (
+        <Skeleton animation="wave" height={300} {...props} />
+    ) : (
+        children
+    );
+};
